Move Eve principal scope adapter under agent
