+++ agent/hooks/session-owner.ts
@@ -2,7 +2,7 @@ import { defineHook } from "eve/hooks";
 import { saveChat } from "@/db/services/chats";
 import { ensureScope } from "@/db/services/scope";
 import { claimSession } from "@/db/services/sessions";
-import { scopeFromPrincipal } from "@/lib/access-scope";
+import { scopeFromPrincipal } from "@/agent/lib/principal-scope";
 
 export default defineHook({
   events: {
