+++ agent/lib/schedules/tools.ts
@@ -1,7 +1,7 @@
 import type { ToolContext } from "eve/tools";
 import { z } from "zod";
 import type { listScheduledAgentJobs } from "@/db/services/scheduled-agent-jobs";
-import { scopeFromPrincipal } from "@/lib/access-scope";
+import { scopeFromPrincipal } from "@/agent/lib/principal-scope";
 
 export function scheduleOwner(context: ToolContext) {
   const auth = context.session.auth.current;
