+++ agent/memory/personal_info.ts
@@ -7,8 +7,9 @@ import {
 } from "eve/memory";
 import { defineTool } from "eve/tools";
 import { z } from "zod";
+import { scopeFromPrincipal } from "@/agent/lib/principal-scope";
 import { readUserProfile, patchUserProfile } from "@/db/services/user-profile";
-import { scopeFromPrincipal, type AccessScope } from "@/lib/access-scope";
+import type { AccessScope } from "@/lib/access-scope";
 import {
   hasUserProfileValues,
   userProfilePatchSchema,
