+++ agent/subagents/worker/hooks/trace-telemetry.ts
@@ -8,7 +8,7 @@ import {
 import { traceTimelineRows } from "@/agent/subagents/worker/lib/trace/timeline";
 import { listWorkerBrowserSessions } from "@/db/services/browsers";
 import type { AccessScope } from "@/lib/access-scope";
-import { scopeFromPrincipal } from "@/lib/access-scope";
+import { scopeFromPrincipal } from "@/agent/lib/principal-scope";
 import { taskCompletionOutputSchema } from "@/lib/worker-completion";
 import { harvestBrowserTraceDomains } from "@/agent/subagents/worker/lib/trace/domains";
