+++ src/lib/access-scope.ts
@@ -1,15 +1,4 @@
 import { createHash } from "node:crypto";
-import type { ConnectionPrincipal } from "eve/connections";
-import type { SessionAuthContext } from "eve/context";
-import { z } from "zod";
-
-const principalScopeSchema = z.object({
-  attributes: z.object({
-    workspaceId: z.string().min(1),
-  }),
-  id: z.string().min(1).optional(),
-  principalId: z.string().min(1).optional(),
-});
 
 export interface AccessScope {
   readonly userId: string;
@@ -27,19 +16,4 @@ export function accessScopeForUser(userId: string): AccessScope {
       .digest("hex")
       .slice(0, 32)}`,
   };
-}
-
-export function scopeFromPrincipal(
-  input: SessionAuthContext | Extract<ConnectionPrincipal, { type: "user" }>
-) {
-  const principal = principalScopeSchema.parse(input);
-  const userId = principal.id ?? principal.principalId;
-  if (!userId) {
-    throw new Error("An authenticated workspace user is required.");
-  }
-
-  return {
-    userId,
-    workspaceId: principal.attributes.workspaceId,
-  } satisfies AccessScope;
 }
